orientation-event/DeviceMotion/screen-upright.html

<!DOCTYPE html>
<html>

  <head>
	<title>02 - A device with the screen upright:</title>
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
	<script src="http://w3c-test.org/resources/WebIDLParser.js"></script>
	<script src="http://w3c-test.org/resources/idlharness.js"></script>
  <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  	<h1>02 - A device with the screen upright:</h1>
	<div id="log">
	</div>
	
	<script>
	
	function init(){
	   if (window.DeviceMotionEvent) {
	   	if(confirm("Put your device with the screen upright.")){
				window.addEventListener('devicemotion', DeviceMotionHandler, false);
			}
		};
	}
	
	function DeviceMotionHandler(eventData){
			
		var gravity = eventData.accelerationIncludingGravity;
		// X : angle par rapport a x
		x = gravity.x;
		// Y : angle par rapport a y
		y = gravity.y;
		// Z : acceleromètre (de base : -9.81)
		z = gravity.z;
		
		tx.step(function() {assert_approx_equals(x, 0, 1)});
		ty.step(function() {assert_approx_equals(y, -10, 1)});
		tz.step(function() {assert_approx_equals(z, 0, 1)});
		tx.done();
		ty.done();
		tz.done();
	}

	init();
	
	var tx = async_test("X angle");
	var ty = async_test("Y angle");
	var tz = async_test("Z angle");
	</script>
	
  </body>
  
</html>
